# TrillionBit-API-stage/TrillionBit-API-stage/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bitex application programming interface (API) allows our clients to access and control their accounts, using custom written software. <br/> With the inclusion of the new BTC/AED currency pair, we've added new endpoints for some API calls (info about specific calls below). The old endpoints are still available and have not changed. We strongly suggest however, that you refer to the API for future references.",
        "version": "1.0.0",
        "title": "Bitex",
        "termsOfService": "http://bitex.com",
        "contact": {
        "email": "admin@bitex.com"
    },
    "license": {
        "name": "Bitex API",
        "url": "http://bitex.com"
    }
    },
    "host": "api.bitex.com",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/trading/order/finished": {
            "post": {
                "tags": [
                    "Trading"
                ],
                "summary": "Get order finished.",
                "description": "Get order finished",
                "operationId": "order.finished",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Get order finished",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "market": {
                                    "type": "string"
                                },
                                "start_time": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "end_time": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "offset": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "limit": {
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "side": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "create:order"
                        ]
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://api.bitex.com/api/auth/login",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
...
}